requirements.txt
aicrowd-cli
python-chess
flask
openai>=1.0.0
jinja2>=3.1.6
# -r chess-env/requirements.txt  # Commented out - clone chess-env repo first if needed